requirements.txt
FastApi~=0.110.0
uvicorn~=0.29.0
sqlalchemy~=2.0.29
psycopg2
python-status

pydantic~=2.6.4